major rearrangement, introduction of Flask app factory, and blueprints for individual modules.
