+++ app/status/__init__.py
@@ -1,0 +1,8 @@
+"""
+The status Blueprint handles the retrieval of status objects
+"""
+from flask import Blueprint
+
+status_blueprint = Blueprint('status', __name__, template_folder='status')
+
+from . import routes
