+++ service/elasticsearch_service.py
@@ -1,0 +1,32 @@
+import json
+
+from elasticsearch import Elasticsearch
+
+from model import AllResponses
+from model.UpdateContainer import UpdateContainer
+
+es = Elasticsearch()
+
+
+def send_to_index(all_responses: AllResponses, query_id):
+    all_responses_json = json.dumps(all_responses, cls=HiddenEncoder)
+    res = es.index(query_id, 'all_data', all_responses_json, all_responses.id, request_timeout=600)
+    print('saved to index ' + query_id)
+    return res
+
+
+def append_to_index(document, eid, query_id):
+    update_container = UpdateContainer(document)
+    update_json = json.dumps(update_container, cls=HiddenEncoder)
+    res = es.update(index=query_id, doc_type="all_data", id=eid, body=update_json)
+    print('saved to index ' + query_id)
+    return res
+
+
+def delete_index(project_id):
+    es.indices.delete(project_id, ignore=[400, 404])
+
+
+class HiddenEncoder(json.JSONEncoder):
+    def default(self, o):
+        return {k.lstrip('_'): v for k, v in o.__getstate__().items()}
