+++ service/relevance_measure_service.py
@@ -1,6 +1,9 @@
 import json
+from json import JSONDecodeError
 
 from flask import current_app as app
+
+from model.RelevanceMeasures import RelevanceMeasure
 
 
 def save_relevance_measures(project_id, relevance_measures):
@@ -8,7 +11,7 @@ def save_relevance_measures(project_id, relevance_measures):
         location = app.config.get("LIBINTEL_DATA_DIR")
     path_to_file = location + '/out/' + project_id + '/relevance_measures.json'
     with open(path_to_file, 'w') as json_file:
-        json_file.write(json.dumps(relevance_measures.__dict__))
+        json_file.write(json.dumps(relevance_measures))
         json_file.close()
 
 
@@ -18,7 +21,13 @@ def load_relevance_measure(project_id):
     path_to_file = location + '/out/' + project_id + '/relevance_measures.json'
     print(path_to_file)
     with open(path_to_file) as json_file:
-        relevance_measure = json.load(json_file)
-        print(relevance_measure)
-        json_file.close()
-        return relevance_measure
+        try:
+            relevance_measure = json.load(json_file)
+            print(relevance_measure)
+            json_file.close()
+            return relevance_measure
+        except FileNotFoundError:
+            return {}
+        except JSONDecodeError:
+            return {}
+
