--- static/css/jumbotron.css
@@ -1,5 +1,0 @@
-/* Move down content because we have a fixed navbar that is 50px tall */
-body {
-  padding-top: 40px;
-  padding-bottom: 10px;
-}
