+++ utilities/HiddenEncoder.py
@@ -1,0 +1,6 @@
+import json
+
+
+class HiddenEncoder(json.JSONEncoder):
+    def default(self, o):
+        return {k.lstrip('_'): v for k, v in o.__getstate__().items()}
